// src/app/domain/enum/contributor-role.enum.ts
export enum ContributorRole {
  DataCollector = 'Data Collector',
  DataCurator = 'Data Curator',
  DataManager = 'Data Manager',
  Distributor = 'Distributor',
  Editor = 'Editor',
  HostingInstitution = 'HostingInstitution',
  Producer = 'Producer',
  ProjectLeader = 'ProjectLeader',
  ProjectManager = 'Project Manager',
  ProjectMember = 'Project Member',
  RegistrationAgency = 'Registration Agency',
  RegistrationAuthority = 'Registration Authority',
  RelatedPerson = 'Related Person',
  Researcher = 'Researcher',
  ResearchGroup = 'Research Group',
  RightsHolder = 'Rights Holder',
  Sponsor = 'Sponsor',
  Supervisor = 'Supervisor',
  WorkPackageLeader = 'Work Package Leader',
  Other = 'Other'
}